+++ src/pagination-controls.component.spec.ts
@@ -70,6 +70,7 @@ describe('PaginationControlsComponent:', () => {
         fixture.detectChanges();
 
         const listItems = fixture.debugElement.queryAll(By.css('pagination-controls li'))
+            .filter(el => (el.nativeElement as HTMLLIElement).classList.contains('small-screen') === false)
             .map((el: DebugElement) => el.nativeElement);
 
         expect(listItems[2].classList.contains('ellipsis')).toBe(true);
@@ -229,7 +230,7 @@ describe('PaginationControlsComponent:', () => {
     it('"autoHide" should be boolean', () => {
         let fixture = TestBed.createComponent(ComponentTestComponent);
         let controlsInstance = fixture.debugElement.query(By.css('pagination-controls')).componentInstance;
-        expect(controlsInstance.directionLinks).toBe(true);
+        expect(controlsInstance.autoHide).toBe(false);
     });
 
     it('"autoHide" should work with non-data-bound values', () => {
@@ -258,6 +259,12 @@ describe('PaginationControlsComponent:', () => {
         fixture.detectChanges();
 
         expect(getPageLinkItems(fixture).length).toBe(1);
+    });
+
+    it('"directionLinks" should be boolean', () => {
+        let fixture = TestBed.createComponent(ComponentTestComponent);
+        let controlsInstance = fixture.debugElement.query(By.css('pagination-controls')).componentInstance;
+        expect(controlsInstance.directionLinks).toBe(true);
     });
 
     it('"directionLinks" should work with non-data-bound values', () => {
@@ -290,6 +297,31 @@ describe('PaginationControlsComponent:', () => {
         expect(items[0]).toContain('1');
         expect(items[items.length - 1]).toContain('10');
     }));
+
+
+    it('"responsive" should work with non-data-bound values', () => {
+        overrideTemplate(ComponentTestComponent, `
+            <ul>
+                <li *ngFor="let item of collection | paginate: config" class="list-item">{{ item }}</li>
+            </ul>
+            <pagination-controls responsive="true" [id]="config.id"></pagination-controls>`);
+        let fixture = TestBed.createComponent(ComponentTestComponent);
+        fixture.detectChanges();
+        const controlsCmp: PaginationControlsComponent = fixture.debugElement
+            .query(By.css('pagination-controls')).componentInstance;
+
+        expect(controlsCmp.responsive).toBe(true);
+    });
+
+    it('"responsive" state should be reflected in default template', () => {
+        let fixture = TestBed.createComponent(ComponentTestComponent);
+        let instance = fixture.componentInstance;
+        instance.responsive = true;
+        fixture.detectChanges();
+
+        const list = fixture.debugElement.query(By.css('ul.ngx-pagination'))
+        expect(list.classes['responsive']).toBe(true);
+    });
 
     describe('custom labels', () => {